requirements.txt
Flask
google-generativeai
Flask-Cors
gunicorn